+++ app/uk/gov/hmrc/gform/gform/PageUpdater.scala
@@ -29,8 +29,6 @@ class PageUpdater[A <: PageMode](page: Page[A], index: Int, baseIds: List[FormCo
   private def expandRemoveItemIf(removeItemIf: RemoveItemIf) = RemoveItemIf(expandBooleanExpr(removeItemIf.booleanExpr))
 
   private def expandValidator(validator: Validator) = ValidatorUpdater(validator, index, baseIds)
-  private def expandValidators(validators: Validators) =
-    ValidatorsUpdater(validators, index, baseIds) //this will be removed by GFORMS-2279
 
   private def expandDataRetrieve(dataRetrieve: DataRetrieve) = DataRetrieveUpdater(dataRetrieve, index, baseIds)
 
@@ -53,8 +51,7 @@ class PageUpdater[A <: PageMode](page: Page[A], index: Int, baseIds: List[FormCo
       fields = page.fields.map { field =>
         new FormComponentUpdater(field, index, baseIds).updatedWithId
       },
-      validator = page.validator.map(expandValidator),
-      validators = page.validators.map(expandValidators),
+      validators = page.validators.map(expandValidator),
       continueLabel = page.continueLabel.map(expandSmartString),
       instruction = page.instruction.map(i => i.copy(name = i.name.map(expandSmartString))),
       dataRetrieve = NonEmptyList.fromList(page.dataRetrieves().map(expandDataRetrieve)),
