--- app/uk/gov/hmrc/gform/sharedmodel/formtemplate/Validators.scala
@@ -1,43 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.gform.sharedmodel.formtemplate
-
-import julienrf.json.derived
-import play.api.libs.json.OFormat
-import uk.gov.hmrc.gform.sharedmodel.SmartString
-//this will be removed by GFORMS-2279
-sealed trait Validators {
-  def errorMessage: SmartString
-}
-
-case object Validators {
-  implicit val format: OFormat[Validators] = derived.oformat()
-}
-
-case class HmrcRosmRegistrationCheckValidator(
-  errorMessage: SmartString,
-  regime: String,
-  utr: FormCtx,
-  postcode: FormCtx
-) extends Validators {
-  val utrFieldId = utr.formComponentId
-  val postcodeFieldId = postcode.formComponentId
-}
-
-object HmrcRosmRegistrationCheckValidator {
-  implicit val format: OFormat[HmrcRosmRegistrationCheckValidator] = derived.oformat()
-}
