+++ it/uk/gov/hmrc/gform/it/sample/FormTemplateSample.scala
@@ -72,7 +72,6 @@ trait FormTemplateSample {
                 )
               ),
               includeIf = None,
-              validator = None,
               validators = None,
               continueLabel = None,
               continueIf = None,
