+++ test/uk/gov/hmrc/gform/controllers/NavigationSpec.scala
@@ -62,7 +62,6 @@ class NavigationSpec extends Spec with FormModelSupport with VariadicFormDataSup
         description = None,
         shortName = None,
         includeIf = includeIf,
-        validator = None,
         validators = None,
         fields = formComponent :: Nil,
         continueLabel = None,
