+++ test/uk/gov/hmrc/gform/gform/StructuredFormDataBuilderSpec.scala
@@ -1068,7 +1068,6 @@ class StructuredFormDataBuilderSpec
         None,
         None,
         None,
-        None,
         fields.toList,
         None,
         None,
@@ -1086,7 +1085,6 @@ class StructuredFormDataBuilderSpec
     Section.RepeatingPage(
       Page(
         toSmartString("repeating-page"),
-        None,
         None,
         None,
         None,
