+++ test/uk/gov/hmrc/gform/graph/DependencyGraphSpec.scala
@@ -47,7 +47,6 @@ class DependencyGraphSpec extends AnyFlatSpecLike with Matchers with FormModelSu
     None,
     None,
     None,
-    None,
     List.empty[FormComponent],
     None,
     None,
