+++ test/uk/gov/hmrc/gform/graph/FormTemplateBuilder.scala
@@ -57,7 +57,6 @@ object FormTemplateBuilder {
     None,
     includeIf,
     None,
-    None,
     formComponents,
     None,
     None,
@@ -140,7 +139,6 @@ object FormTemplateBuilder {
         None,
         None,
         Some(includeIf),
-        None,
         None,
         formComponents,
         None,
@@ -339,7 +337,6 @@ object FormTemplateBuilder {
     None,
     None,
     None,
-    None,
     formComponents,
     None,
     None,
@@ -361,7 +358,6 @@ object FormTemplateBuilder {
       None,
       None,
       Some(includeIf),
-      None,
       None,
       formComponents,
       None,
