+++ test/uk/gov/hmrc/gform/models/FormModelSpec.scala
@@ -90,7 +90,6 @@ class FormModelSpec extends AnyFlatSpecLike with Matchers with FormModelSupport 
         None,
         None,
         None,
-        None,
         expectedFcs,
         None,
         None,
