+++ test/uk/gov/hmrc/gform/sharedmodel/formtemplate/generators/SectionGen.scala
@@ -94,8 +94,7 @@ trait SectionGen {
       caption          <- Gen.option(smartStringGen)
       shortName        <- Gen.option(smartStringGen)
       includeIf        <- Gen.option(IncludeIfGen.includeIfGen)
-      validator        <- Gen.option(ValidatorGen.validatorGen)
-      validators       <- Gen.option(ValidatorsGen.validatorsGen)
+      validators       <- Gen.option(ValidatorGen.validatorGen)
       fields           <- PrimitiveGen.oneOrMoreGen(FormComponentGen.formComponentGen())
       continueLabel    <- Gen.option(smartStringGen)
       continueIf       <- Gen.option(ContinueIfGen.continueIfGen)
@@ -112,7 +111,6 @@ trait SectionGen {
       shortName,
       caption,
       includeIf,
-      validator,
       validators,
       fields.toList,
       continueLabel,
@@ -121,7 +119,7 @@ trait SectionGen {
       presentationHint,
       None,
       confirmation,
-      redirects.map(NonEmptyList.one(_)),
+      redirects.map(NonEmptyList.one),
       None,
       removeItemIf
     )
